rename API URLs file
