+++ frontend/src/components/jobs/result/JobResult.jsx
@@ -5,7 +5,7 @@ import useInterval from "react-use/lib/useInterval";
 import { useParams } from "react-router-dom";
 
 import { Loader } from "@certego/certego-ui";
-import { JOB_BASE_URI } from "../../../constants/api";
+import { JOB_BASE_URI } from "../../../constants/apiURLs";
 import { JobOverview } from "./JobOverview";
 
 import {
