+++ frontend/src/components/jobs/result/bar/comment/CommentResult.jsx
@@ -4,7 +4,7 @@ import useTitle from "react-use/lib/useTitle";
 import { useParams } from "react-router-dom";
 import { Loader } from "@certego/certego-ui";
 
-import { JOB_BASE_URI } from "../../../../../constants/api";
+import { JOB_BASE_URI } from "../../../../../constants/apiURLs";
 
 import CommentOverview from "./CommentOverview";
