+++ frontend/src/stores/useOrganizationStore.jsx
@@ -1,7 +1,7 @@
 import { create } from "zustand";
 import axios from "axios";
 
-import { BASE_URI_ORG, ORG_PLUGIN_DISABLE_URI } from "../constants/api";
+import { BASE_URI_ORG, ORG_PLUGIN_DISABLE_URI } from "../constants/apiURLs";
 
 const useOrganizationStore = create((set, _get) => ({
   loading: false,
